% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHcircular.R
\name{HHMM.lm}
\alias{HHMM.lm}
\title{Circular-linear regression with "HH:MM" outcomes}
\usage{
HHMM.lm(y, X.formula, name = NULL)
}
\arguments{
\item{y}{String vector of times-of-day in "HH:MM" format.}

\item{X.formula}{A model formula specifying the predictors.}

\item{name}{Optional string vector of variable names.}
}
\value{
A string matrix with columns:
\itemize{
\item "": Parameter names
\item "Time (95\%CI)": Parameter estimates and 95\% confidence intervals
\item "p": \emph{P}-values
}
The μ intercept is presented as a time-of-day whereas predictor coefficients
are presented as signed time-of-day differences.
}
\description{
Fits a circular-linear regression¹ model for time-of-day outcomes in
"HH:MM" format.
}
\details{
Wrapper function for \code{WH_HHMM_to_rad()},
\code{WH_rad_to_dHHMM()}, \code{WH_rad_to_HHMM()}, and
\code{WH_reg_circlinear()}.
}
\examples{
y <- c("01:00","01:15","01:30","01:45","02:00",
       "23:00","23:15","23:30","23:45","00:00")
X <- c(rep(0, 5), rep(1, 5))
print(HHMM.mean(y))
print(HHMM.lm(y, ~X))
}
\references{
\enumerate{
\item Fisher, N.I. and Lee, A.J., 1992. Regression models for an angular response.
\emph{Biometrics}, pp. 665–677.
}
}
