% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WHcircular.R
\name{HHMM.mean}
\alias{HHMM.mean}
\title{Circular mean of times-of-day in "HH:MM" format}
\usage{
HHMM.mean(x)
}
\arguments{
\item{x}{String vector of times-of-day in "HH:MM" format.}
}
\value{
The circular mean time-of-day in "HH:MM" string format.
}
\description{
Computes the circular mean of times-of-day in "HH:MM" format,
rounded to the nearest minute.
}
\details{
Wrapper function for \code{WH_HHMM_to_rad()},
\code{WH_rad_mean()}, and \code{WH_rad_to_HHMM()}.
}
\examples{
y <- c("01:00","01:15","01:30","01:45","02:00",
       "23:00","23:15","23:30","23:45","00:00")
print(HHMM.mean(y))
}
