% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_CordeiroPaulaBotter}
\alias{WH_CordeiroPaulaBotter}
\title{Circular ANOVA}
\usage{
WH_CordeiroPaulaBotter(theta, group)
}
\arguments{
\item{theta}{Numeric vector of radians.}

\item{group}{Integer vector of group identifiers.}
}
\value{
\emph{P}-value of the test statistic.
}
\description{
One-way analysis of variance for circular data in radians using the
Cordeiro–Paula–Botter¹ likelihood-ratio test.
}
\examples{
y <- c("01:00","01:15","01:30","01:45","02:00",
       "23:00","23:15","23:30","23:45","00:00")
X <- c(rep(0, 5), rep(1, 5))
theta <- WH_HHMM_to_rad(y)
print(WH_CordeiroPaulaBotter(theta, X))
}
\references{
\enumerate{
\item Cordeiro, G.M., Paula, G.A. and Botter, D.A., 1994. Improved likelihood
ratio tests for dispersion models. \emph{International Statistical Review}, pp. 257–274.
}
}
