% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_HHMM_to_rad}
\alias{WH_HHMM_to_rad}
\title{Convert "HH:MM" time-of-day strings to radians}
\usage{
WH_HHMM_to_rad(input)
}
\arguments{
\item{input}{String vector of times in "HH:MM" format.}
}
\value{
A numeric vector of radians.
}
\description{
Converts times-of-day in "HH:MM" string format (for example, "14:00" for 2:00 PM)
to radians.
}
