% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_int_to_rad}
\alias{WH_int_to_rad}
\title{Convert integer times-of-day to radians}
\usage{
WH_int_to_rad(x)
}
\arguments{
\item{x}{Integer vector of times in HHMM format.}
}
\value{
A numeric vector of radians.
}
\description{
Converts times-of-day in HHMM integer format (for example, 1400 for 2:00 PM)
to radians.
}
