% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_rad_resultant}
\alias{WH_rad_resultant}
\title{Circular resultant vector}
\usage{
WH_rad_resultant(theta)
}
\arguments{
\item{theta}{Numeric vector of radians.}
}
\value{
A numeric vector of 2 elements:
\itemize{
\item Mean direction in radians
\item Resultant length
}
}
\description{
Computes the circular resultant vector in polar coordinates
from a vector of radians. If the resultant length is
smaller than \code{DBL_EPSILON}, the mean direction is set to \code{NaN}.
}
