% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_rad_sd}
\alias{WH_rad_sd}
\title{Circular standard deviation}
\usage{
WH_rad_sd(theta)
}
\arguments{
\item{theta}{Numeric vector of radians.}
}
\value{
The circular standard deviation in radians.
}
\description{
Computes the circular standard deviation of a vector of radians.
}
\examples{
y <- c("01:00","01:15","01:30","01:45","02:00",
       "23:00","23:15","23:30","23:45","00:00")
X <- c(rep(0, 5), rep(1, 5))
theta <- WH_HHMM_to_rad(y)
print(WH_rad_to_HHMM(WH_rad_sd(theta)))
}
