% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_rad_to_dHHMM}
\alias{WH_rad_to_dHHMM}
\title{Convert radians to signed "HH:MM" time differences}
\usage{
WH_rad_to_dHHMM(input)
}
\arguments{
\item{input}{Numeric vector of radians.}
}
\value{
A string vector of signed time-of-day differences in "HH:MM" format.
}
\description{
Converts radians to time-of-day differences in "HH:MM" string format,
rounded to the nearest minute. Negative values are prefixed with a minus sign.
}
