% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_rad_to_HHMM}
\alias{WH_rad_to_HHMM}
\title{Convert radians to "HH:MM" times}
\usage{
WH_rad_to_HHMM(input)
}
\arguments{
\item{input}{Numeric vector of radians.}
}
\value{
A string vector of times-of-day in "HH:MM" format.
}
\description{
Converts radians to times-of-day in "HH:MM" string format,
rounded to the nearest minute.
}
