% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{WH_rad_kappa}
\alias{WH_rad_kappa}
\title{Circular concentration}
\usage{
WH_rad_kappa(theta)
}
\arguments{
\item{theta}{Numeric vector of radians.}
}
\value{
The maximum likelihood estimate of κ.
}
\description{
Computes the von Mises concentration parameter κ of a vector
of radians.
}
